Create cricket score board
Project Name: Cricket score Board

Team Members:
Prashant Shah, Sahil Prasad Rouniyar, Pratima Chaudhary
,Sandhya Shah

Project Discription
This is the project which is made form c language. This project can store the name , runs ,balls , sixes and fours of the players. This prohect can print the score card of the palyers. 
In this program you can add new player, delete,update the new palyers and at last you can see their score card in the terminal.

Instruction To Run the Project
THis is a simple project so user will not face any difficulty to run this code. Firstly the progarm wil ask the number of players whose data you want to input then you have to add the details of the players and after you add their details you will be given some option like enter this number to print score card or add new players etc.

Aditional Information 
This is the first time we are doing a project so it has lot of things that can be improved.I believe that as we do more projects the quality of the projects and it user friendliness will improve.